Don't delete old tarballs if there is no wiki to export.  D'oh
git-svn-id: https://repo.socialtext.net:8999/svn/socialtext/trunk@7582 593ce9bc-2028-0410-bb98-c566b643a948

+++ nlw/share/workspaces/wikitests/appliance/package.bash
@@ -6,23 +6,29 @@
 
 cd  $ST_CURRENT/nlw/share/workspaces/wikitests/appliance
 
-if [ -e $ST_CURRENT/nlw/share/workspaces/wikitests/wikitests.1.tar.gz ]; then
-    echo DELETING old wikitests.1.tar.gz
-    rm $ST_CURRENT/nlw/share/workspaces/wikitests/wikitests.1.tar.gz
+if [ -e ~/.nlw/root/data/wikitests ]; then
+    
+    if [ -e $ST_CURRENT/nlw/share/workspaces/wikitests/wikitests.1.tar.gz ]; then
+        echo DELETING old wikitests.1.tar.gz
+        rm $ST_CURRENT/nlw/share/workspaces/wikitests/wikitests.1.tar.gz
+    fi
+    
+    echo CREATING new wikitests.1.tar.gz
+    $ST_CURRENT/nlw/bin/st-admin export-workspace --w wikitests --dir $ST_CURRENT/nlw/share/workspaces/wikitests/
+    cp $ST_CURRENT/nlw/share/workspaces/wikitests/wikitests.1.tar.gz $ST_CURRENT/nlw/share/workspaces/wikitests/appliance
 fi
 
-echo CREATING new wikitests.1.tar.gz
-$ST_CURRENT/nlw/bin/st-admin export-workspace --w wikitests --dir $ST_CURRENT/nlw/share/workspaces/wikitests/
-cp $ST_CURRENT/nlw/share/workspaces/wikitests/wikitests.1.tar.gz $ST_CURRENT/nlw/share/workspaces/wikitests/appliance
 
-if [ -e $ST_CURRENT/nlw/share/workspaces/calctests/calctests.1.tar.gz ]; then
-    echo DELETING old calctests.1.tar.gz
-    rm $ST_CURRENT/nlw/share/workspaces/calctests/calctests.1.tar.gz
+if [ -e ~/.nlw/root/data/calctests ]; then
+    if [ -e $ST_CURRENT/nlw/share/workspaces/calctests/calctests.1.tar.gz ]; then
+        echo DELETING old calctests.1.tar.gz
+        rm $ST_CURRENT/nlw/share/workspaces/calctests/calctests.1.tar.gz
+    fi
+    
+    echo CREATING new calctests.1.tar.gz
+    $ST_CURRENT/nlw/bin/st-admin export-workspace --w calctests --dir $ST_CURRENT/nlw/share/workspaces/calctests/
+    cp $ST_CURRENT/nlw/share/workspaces/calctests/calctests.1.tar.gz $ST_CURRENT/nlw/share/workspaces/wikitests/appliance
 fi
-
-echo CREATING new calctests.1.tar.gz
-$ST_CURRENT/nlw/bin/st-admin export-workspace --w calctests --dir $ST_CURRENT/nlw/share/workspaces/calctests/
-cp $ST_CURRENT/nlw/share/workspaces/calctests/calctests.1.tar.gz $ST_CURRENT/nlw/share/workspaces/wikitests/appliance
 
 rm -rf $ST_CURRENT/nlw/share/workspaces/wikitests/appliance/usr
